Restrict event expression for pre- and postcondition
The event expression is a non pure expression in pre- and postcondition. This is the case because we assume these expression in some locations.

Therefore, we restrict the usage of events there:
- Only one event expression can be used per pre- or postcondition
- Event expressions cannot be used in expression requiring pure expressions

If we inhale an expression containing an event expression, we first exhale the expression with all permission to the event. So, exactly the amount we assume of an event is present after the inhale.