Enable checks for private functions
When we inlined a private function, which makes an external call, the checks of the public function are checked as well.

When we just assume the private function, we do not check the checks of the public function directly.
For this, checks get enabled for private functions and an implication is added when we assume the private function ("checks of private function" ==> "checks of public function").

A private function can now also state that it has no public state using "check: False".