Introduce interpreted decorator for lemmas
In lemma_assert the interpreted(...) function can be used to get properties about the uninterpreted functions using interpreted ones.
With the new decorator this is also possible for lemmas.
If a lemma is decorated with "interpreted", the whole lemma is asserted using interpreted functions and then assumed with the uninterpreted functions.