Fix wrapped-integer type handling in loops and conditionals
We cannot track if a variable really is constant or not. That is why, we have to assume that an integer variable that got changed in a loop or a conditional has now multiple possible values.
Non constant integer variables should have the wrapped-integer type. Therefore, update the type of all integer variables that got modified in a loop or conditional.

With the exception of loops over a range-expression. For these, we can make a precise analysis if it does contain wrapped-integer information.