Reworked old state handling for inter contract invariants
Inter contract invariants were not sound. The example that is now added as a test could be falsely verified.
In this example, the state before the external function call has an additional information that all possible values of the token are greater or equal to this value (0 is the smallest unsigned integer value).

This shows that we cannot use this state as the old state. Therefore, a proper handling for the old state was introduced. There is now an old state that models all states before the external call creates a public state. And a further old state, to model all the states after and including the last created public state.