Fix issue #21
- Refactor type_annotator.py
- Introduce TypeResolver

A depth-first search in post-order is performed to annotate the type. Previously, the type for each expression is resolved when its type was analysed. But for some expressions like Num or List this cannot be done precisely. Therefore, just the first possibility was selected.

Now, we keep this ambiguity by keeping multiple possible types for expressions and resolve it after the annotation phase.